hello-map: add lighting and occulsion culling
While the light emissions in the texture of the `3D Scifi Kit Starter
Kit` asset do indeed illuminate the prefabs' corridors and rooms, they
results in a lack of shadows.

Prevent the above texture from adding to global illumination, and modify
the light prefabs to include a `Point` light.

Add occulsio culling to prevent rendering any objects outside of the
main camera's view.