hello-map: monitor distance along path
While the Euclidean distance between the player and the NPC can be
useful, the best value that can be used by both a Neural Network and a
Genetic Algorithm is the distance along the checkpoint path.

Make the NPC stop a short distance ahead of the checkpoint, waiting for
the player to reach it before moving on.

If the player is behind by one checkpoint, add the distance between the
player and the checkpoint to the distance of the latter and the NPC. The
result can be used instead of the Euclidean distance between the player
and the NPC.