hello-map: update checkpoint and obstacle system
The previous "distance along path" system was messy, and could be
approached by simply limiting the distance the between the stupid-NPC,
and the player character.

Moreover, a key system is required so that the player can progress once
the stupid-NPC moves past the door (which shuts behind them).

Use a photo frame as a temporary placeholder for "key" items and make
sure that obstacles can be put in front of them, thus blocking the
possibility of "pressing 'E'".

Create a UI Text object showing "Press 'E'" in the middle of the screne.

Cast a ray (head-height) from the player character's `CharacterMover`,
and OR the binary values of the `Environment` and `Key` layers, so that
the ray detects collisions on them. When the ray hits the `Key` layer
only, pass the collider's instance ID into the `KeyTextController`.

`KeyTextController` is a script controlling the UI Text object, which
acts as the "Press 'E'" prompt. It looks for the `PressE` object, and
sets its `active` value, depending on whether `LinkedObject` is not -1.

Check the instance ID of the above `LinkedObject` in the
`CheckpointController` against that of the key assigned to the current
checkpoint the player is on. This ensures that the key only works for
the doors in that location, and not for any others.

Add "back doors", which close behind the player, preventing them from
moving to the previous checkpoints.

Remove the "distance along path" system and, instead, prevent the
stupid-NPC from continuing to move if it is 10 units or further from the
player character.

Create a new orthographic camera that looks up at the map. The floors
don't render when viewed from below, so this is the ideal location.

Render the camera's raw image onto a `Render Texture`, and show this in
a `Raw Image` UI object, on the screen overlay. Generate a green dot for
the player character and red dots for any NPCs.