hello-map: update checkpoint system and obstacles
When the stupid-NPC would cross the door boundaries, and the player
would try opening it, the door would keep shutting.

Use the `met` boolean in `CheckpointController` to set a flag to `true`,
when the stupid-NPC crosses a door, and it closes behind them.

Under the NPC's `CharacterBehaviour`, make sure the door only shuts when
the checkpoint isn't `met`.

Add a few more obstacles.

Different obstacles may need to be used, and the gaps through which the
rays can see the keys may need to be increased; the GA and NN may have a
difficult time opening the door.