hello-map: implement level changes and end trigger
Before going onto create new levels, a level controller and level
"ender" have to be implemented. Some additional touches to the character
objects also need to be made, so that the collisions aren't buggy.

Create a floating, invisible portal in an appropriate location, which
(upon colliding with a player) triggers a scene change. The next scene
should be that which is next in queue, under "Build Settings -> Scenes
in Build".

Create an empty placeholder scene, which is loaded either as the last
scene in the queue (by the level portal), or when a player collides with
the stupid-NPC.

In the future, depending on whether the portal or stupid-NPC has been
used to load `LevelEnd`, the game will have either been failed or lost.
This result is irrelevant to the overall experiment, but could help with
the user experience, when being played by a group of random players (as
part of the training data collection).