hello-map: use character template in level changer
While the existing game implementation worked, it made it difficult to
make changes to objects that repeat over all levels, namely the player
and stupid-NPC characters. Unity allows to stick a `DontDestroy` flag on
objects, so it is definitely possible to reuse the character objects
throughout the entire game.

Remove all characters from all existing levels, as well as any other
object that should repeat throughout all levels (listed below). From
now, for all levels, create an empty `LevelCoordinates` object that is a
parent to all other objects in the level. The object should have a new
script, `LevelCoordinates` assigned to it.

Create a `LevelZero` scene, which shows a button and contains objects
that repeat for all levels:
- `MultilevelController` - empty container for the new
`MultilevelController` script
- `AllCharacters` - the parent container that holds both the `Player`
character container and that of `NPC`
- `Debug Text Canvas`
- `KeyPrompt`

Create a script that sets a `DontDestroy` flag on any objects that use
it. Allow for setting the "active" status of a `GameObject` in the
script, and apply it after setting `DontDestroyOnLoad`.

Create a new `CheckpointController2` script. This will now be assigned
to the new `AllCheckpoints` container. The container should be a parent
to all checkpoints objects.

From now, all checkpoints should be named by their index (from 1), and
should only contain the following (named exactly as listed):
- `Flag` - previously `flag_med`
- `Door_Front` - previously `Door`
- `Door_Back` - previously `Door`, under a different container,
`BackDoors`
- `Key` - previously `Painting`

Furthermore, only one back door can be set per checkpoint. This
simplifies the code by preventing the need for arrays of doors.

In `CheckpointController2`, find all checkpoints, filter them by their
names (listed above), and encapsulate them in a new `Checkpoint` object.

Declare an array of `Checkpoint` objects in `CharacterBehaviour`,
instead of the previously predefined (in the Scene Editor inspector)
`_allCheckpoints` array.

In `CharacterBehaviour`, remove `_allCheckpointTriggers` and access the
relevant attributes from the one (now public) array, `allCheckpoints`.

Since only the stupid-NPC keeps track of checkpoints, remove
`GlobalCharacter` usage and count `_checkpoint` in `CharacterBehaviour`.

Rename the `NonPlayerCharacter` enum to `NPC`, to match the tags and
layers. Make sure there are tags and layers for both `Player` and `NPC`.

In `CharacterMoverController`, set the `GameObject` tag and layer to
that of its parent container's enum (either `Player` or `NPC`). Ignore
any collisions from `GameObjects` with that tag. This may be useful when
spawning multiple parallel characters (e.g. for the upcoming Genetic
Algorithm).

In `MinimapController`, find all characters (regardless of type) and
assign a "dot", based on their `characterType`.

In the `LevelCoordinates` script, make the spawn position and rotation
(for both the player and NPC) show up in the inspector, to be predefined
in the Scene Editor. Create a `ConfigureLevel` method, to be manually
called from `MultilevelController`.

The level can only be configured once the scene has loaded. Without the
manual call, `LevelCoordinates` is executed before
`MultilevelController`, therefore `allCharacters` stays uninitialised.

To configure the level, give `CharacterBehaviour` all checkpoints under
the `AllCheckpoints` container, move all characters to their predefined
positions (for that level), and add them to the minimap.

Even though the model still doesn't allow for multiple sets of
characters to roam the levels asynchronously (because of doors opening
for-all instead of for-each character), it is not possible to spawn
characters at locations predefined per level.

Changes to characters and some other objects can now be made from
`LevelZero`. The code has also been slightly cleaned, but there are
still a lot of code smells left. This isn't as important as implementing
the Genetic Algorithm and Neural Network, so it won't be a focus in the
nearby future.