nn: create final model and cleanup
For the purposes of this project, the "final" model has been created.
While it doesn't work as effectively as was hoped, interestingly, it had
a 64% success rate during a Turing Test. Weirdly, a lot of participants
thought the network was actually a human being playing the game, and
vice versa.

To put things simple: the data-loading method was changed from a for
loop, to a file crawler, the neural net was changed, and some custom
arithmetic functions had to be used.